requirements.txt
# ==============================================================================
# File: requirements.txt
# Location: root of your project directory
# Purpose: Lists all project dependencies.
# ==============================================================================
Flask==2.3.3
yfinance==0.2.37
pandas==2.2.2
gunicorn==21.2.0
matplotlib==3.8.4
numpy==1.26.4
scipy==1.13.1
requests==2.31.0
requests-cache==1.2.0
Jinja2==3.1.2 
python-dotenv==1.1.1

